+++ packages/utils/src/lib/collect/implementation/mock/schema-helper.mock.spec.ts
@@ -4,13 +4,14 @@ import {
   pluginConfigSchema,
   reportSchema,
 } from '@quality-metrics/models';
+import {dummyConfig, dummyReport} from "./config-and-report-dummy.mock";
 import {
-  dummyConfig,
-  dummyReport,
   nxValidatorsOnlyConfig,
   nxValidatorsOnlyReport,
-  nxValidatorsPlugin,
-} from './config-and-report.mock';
+  nxValidatorsPlugin
+} from "./config-and-report-nx-validators.mock";
+import {lighthouseConfig, lighthousePlugin, lighthouseReport} from "./config-and-report-lighthouse.mock";
+
 
 // @NOTICE ATM the data structure changes a lot so this test is a temporarily check to see if the dummy data are correct
 describe('dummy data', () => {
@@ -20,11 +21,19 @@ describe('dummy data', () => {
   });
 });
 
-// @NOTICE ATM the data structure changes a lot so this test is a temporarily check to see if the dummy data are correct
 describe('Nx Validators data', () => {
   it('is valid', () => {
     expect(() => pluginConfigSchema.parse(nxValidatorsPlugin())).not.toThrow();
     expect(() => coreConfigSchema.parse(nxValidatorsOnlyConfig)).not.toThrow();
     expect(() => reportSchema.parse(nxValidatorsOnlyReport)).not.toThrow();
+  });
+});
+
+
+describe('lighthouse data', () => {
+  it('is valid', () => {
+     expect(() => pluginConfigSchema.parse(lighthousePlugin())).not.toThrow();
+    expect(() => coreConfigSchema.parse(lighthouseConfig)).not.toThrow();
+     expect(() => reportSchema.parse(lighthouseReport)).not.toThrow();
   });
 });
