+++ packages/utils/src/lib/collect/implementation/utils.ts
@@ -1,4 +1,7 @@
 import { CategoryConfig } from '@quality-metrics/models';
+import {dirname, join} from "path";
+import {fileURLToPath} from "url";
+import {readFileSync} from "fs";
 
 export const reportHeadlineText = 'Code Pushup Report';
 export const reportOverviewTableHeaders = ['Category', 'Score', 'Audits'];
@@ -13,3 +16,37 @@ export function countWeightedRefs(refs: CategoryConfig['refs']) {
     .filter(({ weight }) => weight > 0)
     .reduce((sum, { weight }) => sum + weight, refs.length);
 }
+
+
+export class ReadPackageJsonError extends Error {
+  constructor(message: string) {
+    super(`error reading package.json: ${message}`);
+  }
+}
+
+export async function readPackageJson() {
+  try {
+    const __dirname = fileURLToPath(dirname(import.meta.url));
+    const filepath = join(
+      __dirname,
+      '..',
+      '..',
+      '..',
+      '..',
+      '..',
+      'cli',
+      'package.json',
+    );
+    console.log('load ' + filepath);
+
+    return JSON.parse(readFileSync(filepath).toString()) as {
+      name: string;
+      version: string;
+    };
+  } catch (e) {
+    const _e = (e as { message: string });
+    console.warn(_e.message)
+    throw new ReadPackageJsonError(_e.message);
+  }
+}
+
